Fixed issues and started styling
Fixed all remaining issues with toggling of Search Bar Component. Begun
to format and style components in the app. Dropdown component fully
styled for appMenu configuration.

After fixing the main issue with the SearchBar component's container
div only spanning a certain area, the component's menu can now be
clicked out of from anywhere in the page, not just the the div.
With all the components functioning properly the last thing to do
is to put them into place on the page and style them for proper
presentation. Already the DropDown component has been fully styled
to fit the main app menu configuration, so it looks like a proper
menu.