continued to style elements at the top of the page.
Made some minor tweaks to the drop-down component and
consolidated its css file. Moved on to the search bar component
and the main logo of the page.

After fixing some issues with the sizing and centering of the image
within the drop-down app menu, and applying some new colors and
animations, I have begun to style the next component that will be placed
right next to the app menu, the search bar. I had to create a new
container div for all the elements that go up in the top bar of the page
so that I can resize and realign them easier. This made for some issues
with the drop-down I had already put in because I had custom sized it
so it wasn't inheriting any sizes from the container div. Now any
element within the container div is styled to match it's size through
flexbox. The next step is to decide where to align these elements along
the top row such that they look pleasant. I am encountering issues with
aligning certain elements in the center of the screen, but I will
address that later

Signed off by: Robert Yaron <rykanohi1@hotmail.com>