Finished styling SearchBar, new components needed.
Styling for the SearchBar component is complete, and most of the
navigation bar is done as well. My DropDown component does not cover
speific situations where something other than lists of buttons are
required. I will have to make some new components to fit these niches.

	After finishing the css on my SearchBar component and redoing some
styling on the DropDown component, I moved on to the rest of the
navigation bar. I am fairly pleased with the look so far and all that
was left for the most part is to add in some buttons on the end and style
them.
	However I discovered that the navigation bar that I am modeling my
own off of does not just use buttons, it uses drop-down menus that are
different from the DropDown component I made. These drop-downs contain
search bars, text elements, check boxes, as well as buttons. My current
DropDown component does not cover this wide a range of customization,
and I believe it would be overly complicated to make one DropDown
component that you could put anything into and have it work. So instead
I will be making specialized components to fit the roles that my
DropDown component cannot. This has been quite a setback, especially
since all my code alterations have broken most of my tests, so they will
need to be redone as well.

Signed off by: Robert Yaron <rykanohi1@hotmail.com>