Made first test for new component
Fixed most of the broken test and made new test suite for the first of
my new components, as well as the first test.

Most of the broken tests have been fixed, a few are still causing issues
but those will be fixed later. Next I made the new test file for my
first new component, the settings component, and made the first test.
there are some issues concerning the state management of the component
that will have to be addressed.

Signed off by: Robert Yaron <rykanohi1@hotmail.com>